Tried fixing pool problem with distirbution
There is a problem where if you stop the code during distributed runs, the processes that got submitted will still try to run even though the main process has stopped. You cannot easily stop these processes. I have tried working on some code to fix that, but it does not work as well as it should.